Split out lsp-zero config
