+++ neovim/nvim/lua/custom/plugins/treesitter.lua
@@ -46,11 +46,11 @@ return {
         "python",
         "query",
         "regex",
-        "toml",
+        -- "toml",
         -- "tsx",
         "typescript",
-        -- "vim",
-        -- "vimdoc",
+        "vim",
+        "vimdoc",
         "yaml",
         "go",
         "c_sharp",
