+++ neovim/nvim/plugin/options.lua
@@ -23,7 +23,7 @@ vim.opt.hlsearch = false
 vim.opt.incsearch = true
 vim.opt.showmatch = false -- This was annoying
 
-vim.opt.scrolloff = 10 -- ALways have 10 lines below the cursor
+vim.opt.scrolloff = 10    -- ALways have 10 lines below the cursor
 vim.opt.signcolumn = "yes"
 vim.opt.isfname:append("@-@")
 vim.opt.updatetime = 1000    -- Update window faster
